Reorganize remaining bits of common.ts
