--- .changeset/seven-taxis-sort.md
@@ -1,5 +1,0 @@
----
-"create-cloudflare": patch
----
-
-Update SolidStart template to work with create-solid@0.5.3
