+++ packages/create-cloudflare/templates/solid/c3.ts
@@ -64,7 +64,7 @@ const configure = async (ctx: C3Context) => {
 const config: TemplateConfig = {
 	configVersion: 1,
 	id: "solid",
-	displayName: "SolidStart",
+	displayName: "Solid",
 	platform: "pages",
 	generate,
 	configure,
